+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 All notable changes to this project will be documented in this file.
 
+## Unreleased
+
+* Another round of improvements to experimental vector tile rendering.
+* If raster image cannot be decoded, Walkers will try to interpret it as a vector tile. This means
+  that raster and mvt are supported in both PmTiles and HTTP tile sources.
+
 ## 0.49.0
 
 * More improvements to experimental vector tile rendering.